refactor: split orders.service.ts into orders/refunds/production services #327
